coordinator pronto

+++ ProjetoIdiomas/Modules/ViewPost/Views/PostCell.swift
@@ -49,14 +49,14 @@ class PostCell: UITableViewCell {
     @IBAction func downVote(_ sender: Any) {
         downVoted()
         ///atualizo label dos numeros
-        if let tex = downVoteLabel.text, let num = Int16(tex){
+        if let tex = downVoteLabel.text, let num = Int32(tex){
             downVoteLabel.text = "\(num + 1)"
         }
     }
     @IBAction func upVote(_ sender: Any) {
         upvoted()
         ///atualizo label dos numeros
-        if let tex = upVoteLabel.text, let num = Int16(tex){
+        if let tex = upVoteLabel.text, let num = Int32(tex){
             upVoteLabel.text = "\(num + 1)"
         }
     }
